Fixed issues with source code (restarting Xcode/mac), also made the showing of an image optional depending on if there is one in the source file or not
