--- iOS/ContentView.swift
@@ -1,21 +1,0 @@
-//
-//  ContentView.swift
-//  iOS
-//
-//  Created by Moritz Asche on 2022-06-01.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        Text("Hello, world!")
-            .padding()
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
