+++ iOS/Views/NodeView.swift
@@ -14,10 +14,6 @@ struct NodeView: View {
     @Binding var activeNode: Int
     
     // MARK: Computed properties
-    var image: String {
-        return node.image ?? ""
-    }
-    
     var body: some View {
         
         ScrollView {
@@ -35,9 +31,11 @@ struct NodeView: View {
                 }
                 
                 // Show the image if there is one
-                Image(node.image!)
-                    .resizable()
-                    .scaledToFit()
+                if let image = node.image {
+                    Image(image)
+                        .resizable()
+                        .scaledToFit()
+                    }
                 
                 // Show choices, when they exist
                 ForEach(node.edges, id: \.self) { currentEdge in
